# clinical-mdr-api/consumer_api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenStudyBuilder Consumer API",
...
    "version": "0.1.112"
  },
  "paths": {
...
    "/v1/studies": {
      "get": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Get Studies",
        "description": "Returns a paginated list of studies, sorted by the specified sort criteria and order.\n\nEach returned study contains a full list of corresponding study versions, sorted by version start date in descending order.\n\nReturned `version_number` value can be used in other endpoints to retrieve study entities (e.g. visits, activities, etc.)\nassociated with a specific study version.",
        "operationId": "get_studies_v1_studies_get",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          },
          {
            "BearerJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortByStudies",
              "default": "uid"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of items to be returned per page. Must be between 1 and the maximum allowed page size.",
              "default": 10,
              "title": "Page Size"
            },
            "description": "Number of items to be returned per page. Must be between 1 and the maximum allowed page size."
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000000,
              "minimum": 1,
              "description": "Page number of the returned list of entities. Must be between 1 and the maximum allowed page number.",
              "default": 1,
              "title": "Page Number"
            },
            "description": "Page number of the returned list of entities. Must be between 1 and the maximum allowed page number."
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by study ID (case-insensitive partial match), for example `NN1234-5678`.",
              "title": "Id"
            },
            "description": "Filter by study ID (case-insensitive partial match), for example `NN1234-5678`."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Study_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/studies/{uid}/study-visits": {
      "get": {
        "tags": [
          "[V1] Studies"
        ],
        "summary": "Get Study Visits",
        "description": "Returns a paginated list of study visits, sorted by the specified sort criteria and order.\n\nIf `study_version_number` query parameter is provided, study visits\nassociated with the specified study version will be returned.\nOtherwise, visits for the latest study version will be returned.",
        "operationId": "get_study_visits_v1_studies__uid__study_visits_get",
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          },
          {
            "BearerJwtAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Study UID",
              "title": "Uid"
            },
            "description": "Study UID"
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortByStudyVisits",
              "default": "unique_visit_number"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "asc"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Number of items to be returned per page. Must be between 1 and the maximum allowed page size.",
              "default": 100,
              "title": "Page Size"
            },
            "description": "Number of items to be returned per page. Must be between 1 and the maximum allowed page size."
          },
          {
            "name": "page_number",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000000000,
              "minimum": 1,
              "description": "Page number of the returned list of entities. Must be between 1 and the maximum allowed page number.",
              "default": 1,
              "title": "Page Number"
            },
            "description": "Page number of the returned list of entities. Must be between 1 and the maximum allowed page number."
          },
          {
            "name": "study_version_number",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Study Version Number",
              "title": "Study Version Number"
            },
            "description": "Study Version Number",
            "example": "2.1"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponseWithStudyVersion_StudyVisit_"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ValidationDetail"
            },
            "type": "array",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "time",
          "path",
          "method",
          "type",
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "PaginatedResponseWithStudyVersion_StudyVisit_": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self",
            "description": "Pagination link pointing to the current page"
          },
          "prev": {
            "type": "string",
            "title": "Prev",
            "description": "Pagination link pointing to the previous page"
          },
          "next": {
            "type": "string",
            "title": "Next",
            "description": "Pagination link pointing to the next page"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyVisit"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items"
          },
          "study_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StudyVersionSimple"
              },
              {
                "type": "null"
              }
            ],
            "description": "Study version information"
          }
        },
        "type": "object",
        "required": [
          "self",
          "prev",
          "next",
          "items"
        ],
        "title": "PaginatedResponseWithStudyVersion[StudyVisit]"
      },
...
      "PaginatedResponse_Study_": {
        "properties": {
          "self": {
            "type": "string",
            "title": "Self",
            "description": "Pagination link pointing to the current page"
          },
          "prev": {
            "type": "string",
            "title": "Prev",
            "description": "Pagination link pointing to the previous page"
          },
          "next": {
            "type": "string",
            "title": "Next",
            "description": "Pagination link pointing to the next page"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Study"
            },
            "type": "array",
            "title": "Items",
            "description": "List of items"
          }
        },
        "type": "object",
        "required": [
          "self",
          "prev",
          "next",
          "items"
        ],
        "title": "PaginatedResponse[Study]"
      },
...
      "SortByStudies": {
        "type": "string",
        "enum": [
          "uid",
          "id_prefix",
          "number"
        ],
        "title": "SortByStudies"
      },
...
      "SortByStudyVisits": {
        "type": "string",
        "enum": [
          "uid",
          "visit_name",
          "unique_visit_number"
        ],
        "title": "SortByStudyVisits"
      },
      "SortOrder": {
        "type": "string",
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder"
      },
      "Study": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid",
            "description": "Study UID"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Study ID"
          },
          "id_prefix": {
            "type": "string",
            "title": "Id Prefix",
            "description": "Study ID prefix"
          },
          "number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Number",
            "description": "Study number",
            "nullable": true
          },
          "acronym": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acronym",
            "description": "Study acronym",
            "nullable": true
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/StudyVersion"
            },
            "type": "array",
            "title": "Versions",
            "description": "Study versions"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "id",
          "id_prefix",
          "versions"
        ],
        "title": "Study"
      },
...
    },
    "securitySchemes": {
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "api:///API.call": "Make calls to the API"
            },
            "authorizationUrl": "",
            "tokenUrl": ""
          }
        }
      },
      "BearerJwtAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "in": "header",
        "name": "Authorization",
        "description": "Access token that will be sent as `Authorization: Bearer {token}` header in all requests"
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}